python-backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
msal==1.28.0
requests==2.31.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0

